Create README.md
Dedicated and driven professional with a background in civil engineering and a passion for data science. With two years of experience in the construction industry, I have honed my skills in project management and problem-solving. I am actively pursuing opportunities in the data science domain and am excited to contribute my diverse skill set to innovative projects and collaborate with teams dedicated to making impactful insights from data.